Overhaul
A massive overhaul of the db.rb function to allow us to hotswap storage methods at will.

E.g. we will be able to make a db_redis or db_mongodb module and prepend that instead , set the relevant variables and it will run with just changing 3-5 lines of code.